add a little landing page spice
