+++ app/globals.css
@@ -75,7 +75,9 @@
   }
 }
 
-/* Typography */
+/* Typography 
+  - Creds to Jorge Sader: https://github.com/shadcn-ui/ui/discussions/1527#discussioncomment-9396350
+*/
 h1 {
   @apply scroll-m-20 text-4xl font-extrabold tracking-tight lg:text-5xl;
 }
